Add YAML exports.
Following the improve-when-simpler principle I've been using so far,
this slightly changes the format of the YAML document. Instead of
the YAML collections named as such:
`FILES_GENERATED_BY_sct_get_centerline_FILE_KIND_UNKNOWN`

we now put them all under FILE_KIND_UNKNOWN and add a comment with
the specific command that created the file.